Moved definition of mixnet contract common items to separate module
