+++ common/mixnet-contract/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "mixnet-contract"
+version = "0.1.0"
+authors = ["Jędrzej Stuczyński <jedrzej.stuczynski@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+cosmwasm-std = { version = "0.13.2" }
+serde = { version = "1.0", features = ["derive"] }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+schemars = "0.7"
