+++ contracts/mixnet/Cargo.toml
@@ -33,6 +33,7 @@ overflow-checks = true
 backtraces = ["cosmwasm-std/backtraces"]
 
 [dependencies]
+mixnet-contract = { path = "../../common/mixnet-contract" }
 cosmwasm-std = { version = "0.13.2", features = ["iterator"] }
 cosmwasm-storage = { version = "0.13.2", features = ["iterator"] }
 schemars = "0.7"
