+++ contracts/mixnet/src/msg.rs
@@ -1,8 +1,7 @@
 use cosmwasm_std::HumanAddr;
+use mixnet_contract::{Gateway, MixNode};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
-
-use crate::state::{Gateway, MixNode};
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct InitMsg {}
