+++ contracts/mixnet/src/state.rs
@@ -1,12 +1,11 @@
-use cosmwasm_std::Coin;
 use cosmwasm_std::{HumanAddr, Storage};
 use cosmwasm_storage::{
     bucket, bucket_read, singleton, singleton_read, Bucket, ReadonlyBucket, ReadonlySingleton,
     Singleton,
 };
+use mixnet_contract::{GatewayBond, MixNodeBond};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
-use std::fmt::Display;
 
 // Contract-level stuff
 
@@ -26,33 +25,7 @@ pub fn config_read(storage: &dyn Storage) -> ReadonlySingleton<State> {
 }
 
 // Mixnode-related stuff
-
 pub const PREFIX_MIXNODES: &[u8] = b"mixnodes";
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize, JsonSchema)]
-pub struct MixNodeBond {
-    pub(crate) amount: Vec<Coin>,
-    pub(crate) owner: HumanAddr,
-    pub(crate) mix_node: MixNode,
-}
-
-impl Display for MixNodeBond {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        // Write strictly the first element into the supplied output
-        // stream: `f`. Returns `fmt::Result` which indicates whether the
-        // operation succeeded or failed. Note that `write!` uses syntax which
-        // is very similar to `println!`.
-        write!(f, "amount: {:?}, owner: {}", self.amount, self.owner)
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize, JsonSchema)]
-pub struct MixNode {
-    pub(crate) host: String,
-    pub(crate) layer: u64,
-    pub(crate) location: String,
-    pub(crate) sphinx_key: String,
-    pub(crate) version: String,
-}
 
 pub fn mixnodes(storage: &mut dyn Storage) -> Bucket<MixNodeBond> {
     bucket(storage, PREFIX_MIXNODES)
@@ -65,50 +38,6 @@ pub fn mixnodes_read(storage: &dyn Storage) -> ReadonlyBucket<MixNodeBond> {
 // Gateway-related stuff
 
 pub const PREFIX_GATEWAYS: &[u8] = b"gateways";
-
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize, JsonSchema)]
-pub struct GatewayBond {
-    pub(crate) amount: Vec<Coin>,
-    pub(crate) owner: HumanAddr,
-    pub(crate) gateway: Gateway,
-}
-
-impl GatewayBond {
-    pub fn new(amount: Vec<Coin>, owner: HumanAddr, gateway: Gateway) -> Self {
-        GatewayBond {
-            amount,
-            owner,
-            gateway,
-        }
-    }
-}
-
-impl Display for GatewayBond {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        if self.amount.len() != 1 {
-            write!(f, "amount: {:?}, owner: {}", self.amount, self.owner)
-        } else {
-            write!(
-                f,
-                "amount: {} {}, owner: {}",
-                self.amount[0].amount, self.amount[0].denom, self.owner
-            )
-        }
-    }
-}
-
-// TODO: I'm not entirely sure what's up with clippy::field_reassign_with_default
-// since it's not present in almost identical MixNode struct
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize, JsonSchema)]
-pub struct Gateway {
-    pub(crate) mix_host: String,
-    pub(crate) clients_host: String,
-    pub(crate) location: String,
-    pub(crate) sphinx_key: String,
-    /// Base58 encoded ed25519 EdDSA public key of the gateway used to derive shared keys with clients
-    pub(crate) identity_key: String,
-    pub(crate) version: String,
-}
 
 pub fn gateways(storage: &mut dyn Storage) -> Bucket<GatewayBond> {
     bucket(storage, PREFIX_GATEWAYS)
